% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{calc_EM_data}
\alias{calc_EM_data}
\title{data frame calculation}
\usage{
calc_EM_data(d1 = 0.01, d2 = 10, step = 0.1, data)
}
\arguments{
\item{d1}{the starting thickness (mm). default: 0.01 mm}

\item{d2}{the ending thickness (mm). default: 10 mm}

\item{step}{the step that used to calculate the data. default: 0.1 mm. Warning: a step less than 0.1 mm would greatly increase the calculation time.}

\item{data}{a data frame that need to be calculated}
}
\value{
a dataframe containing the calculated reflections loss (dB) as functions of frequency and thickness
}
\description{
Convert the raw data of complex permeability and permittivity into a data frame with calculated reflection loss as functions of frequency and thickness
}
\examples{
calc_EM_data (d1 = 1, d2 = 10, step = 0.5, data = temp)
}
