% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{EM_bandwidth}
\alias{EM_bandwidth}
\title{bandwidth calculation}
\usage{
EM_bandwidth(d1 = 0.01, d2 = 10, RL = -10, data)
}
\arguments{
\item{d1}{the starting thickness, default 0.01 mm}

\item{d2}{the end thickness, default 10 mm}

\item{RL}{the RL that used to calculate the bandwidth, default -10 dB}

\item{data}{the data that need to be calculated}
}
\value{
a data frame with thickness and corresponding bandwidth
}
\description{
calculate the bandwidth of RL< -*dB at the assigned thickness range
}
\examples{
EM_bandwidth(d1=1,d2=5,RL=05,data=temp)
}
