% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{EM_maxRL}
\alias{EM_maxRL}
\title{maximum RL}
\usage{
EM_maxRL(d1 = 0.01, d2 = 10, data)
}
\arguments{
\item{d1}{the starting thickness, default 0.01 mm}

\item{d2}{the end thickness, default 10 mm}

\item{data}{the data that need to be calculated}
}
\value{
a data frame with maximum RL value at each thickness
}
\description{
calculate the maximum RL
}
\examples{
EM_maxRL(d1=1,d2=5,data=temp)
}
